requirements.txt
flask==2.1.2
flask_bootstrap==3.3.7.1
diff-match-patch==20200713
Pillow==9.1.0
opencv-python==4.5.5.64
python-decouple==3.6
WTForms==3.0.1
Flask-WTF==1.0.1
numpy==1.22.3
pydantic==1.9.0
pytesseract==0.3.9
tesseract==0.1.3
imutils==0.5.4
scikit-image==0.19.2
matplotlib==3.5.2